Extract routes from api module
