+++ lib/api/player/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+import state from '../state';
+
+const get = function *() {
+  this.body = JSON.stringify(state.player);
+};
+
+export default get;
