+++ lib/api/state.js
@@ -1,0 +1,23 @@
+'use strict';
+
+export default {
+  player: {
+    pos: {
+      x: 5,
+      y: 5
+    },
+    speed: 1
+  },
+  log: [
+    {
+      id: 1,
+      text: 'Welcome to Snowblind'
+    }
+  ],
+  history: [
+    {
+      x: 4,
+      y: 5
+    }
+  ]
+};
