+++ lib/server/index.js
@@ -7,9 +7,9 @@ import serve from 'koa-static';
 import mount from 'koa-mount';
 import favicon from 'koa-favicon';
 import log from '../logging/index';
-import api from './api';
+import api from '../api/index';
 import startApp from './startApp';
-import appConfig from './../../config/app';
+import appConfig from '../../config/app';
 
 let app = koa();
